# backend/openapi_connector.yaml
openapi: 3.0.0
info:
  title: DAOM 문서 추출 커넥터
  description: |
    AI 기반 문서 데이터 추출 서비스 - Power Automate Custom Connector
    
    ## 주요 기능
    - 📄 문서 업로드 및 데이터 추출
    - 🔍 추출 결과 조회
    - ⏳ 비동기 폴링 지원
    - 📋 모델 목록 조회
    - ❌ 작업 취소
  version: 1.0.0
  contact:
    name: DAOM Support
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: your-daom-instance.azurewebsites.net
        description: DAOM Backend Host
securityDefinitions:
  oauth2_auth:
    type: oauth2
    flow: accessCode
    authorizationUrl: https://login.microsoftonline.com/{tenant-id}/oauth2/v2.0/authorize
    tokenUrl: https://login.microsoftonline.com/{tenant-id}/oauth2/v2.0/token
    scopes:
      api://{client-id}/.default: Default scope
security:
  - oauth2_auth: []
paths:
...
  /connectors/result/{job_id}:
    get:
      operationId: GetExtractionResult
      summary: 🔍 추출 결과 조회
      description: Job ID로 추출 상태 및 결과를 조회합니다
      parameters:
        - name: job_id
          in: path
          type: string
          required: true
          description: 작업 ID
      responses:
        200:
          description: 완료된 결과
          schema:
            $ref: '#/definitions/ExtractionResult'
        202:
          description: 처리 중
          schema:
            $ref: '#/definitions/ExtractionResult'
...
definitions:
  ExtractionResult:
    type: object
    properties:
      job_id:
        type: string
      status:
        type: string
        enum: [pending, processing, completed, failed, cancelled]
      model_id:
        type: string
      model_name:
        type: string
      filename:
        type: string
      extracted_data:
        type: object
        description: 추출된 데이터 (필드 별 값)
      confidence:
        type: number
        description: 평균 신뢰도 (0.0 ~ 1.0)
      error:
        type: string
      metadata:
        type: object
        description: 사용자 정의 메타데이터 (passthrough)
      created_at:
        type: string
...